# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Radware Vulnerable E-commerce API",
    "version": "1.2.0",
    "description": "An intentionally vulnerable e-commerce API designed to demonstrate business logic attacks, focusing on path and query parameters. This version includes Protected Entity Mode which blocks destructive actions on certain demo entities.\n"
  },
...
  "paths": {
...
    "/api/users/{userId}/orders": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "List all orders for a specific user (BOLA Target)",
        "description": "Lists orders. BOLA: Authenticated, but no check if current_user.user_id matches path user_id.",
        "operationId": "listUserOrders",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID of the user whose orders to retrieve (BOLA: can be changed to another user's ID)"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of orders with their items",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Order"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized (Not authenticated)"
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "post": {
        "tags": [
          "Orders"
        ],
        "summary": "Create a new order for a user (BOLA Target)",
        "description": "Creates an order. Products and quantities are specified as dynamic query parameters: `product_id_1`, `quantity_1`, `product_id_2`, `quantity_2`, etc. Vulnerability: `userId` in path can be for another user. `address_id` and `credit_card_id` in query can also belong to another user.\n",
        "operationId": "createUserOrder",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID of the user placing the order (BOLA: can be changed)"
          },
          {
            "name": "address_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Address ID for shipping (Vulnerability: can be another user's address)"
          },
          {
            "name": "credit_card_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Credit Card ID for payment (Vulnerability: can be another user's card)"
          },
          {
            "name": "product_id_1",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID of the first product in the order"
          },
          {
            "name": "quantity_1",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Quantity of the first product"
          },
          {
            "name": "product_id_2",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID of the second product in the order (optional)"
          },
          {
            "name": "quantity_2",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Quantity of the second product (optional)"
          }
        ],
        "responses": {
          "201": {
            "description": "Order created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input (e.g., product not found, insufficient stock, invalid parameters)"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User, Address, or CreditCard not found"
          }
        }
      }
    },
    "/api/users/{userId}/orders/{orderId}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Get a specific order for a user (BOLA Target)",
        "description": "Gets order details. BOLA: Authenticated, but no check if current_user.user_id matches path user_id.",
        "operationId": "getUserOrderById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID of the user (BOLA: can be changed)"
          },
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID of the order to retrieve"
          }
        ],
        "responses": {
          "200": {
            "description": "Detailed information about the order",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User or Order not found"
          }
        }
      }
    },
    "/api/users/{userId}/orders/{orderId}/status-legacy": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Get legacy order status (Deprecated, BOLA Target)",
        "description": "Legacy order status lookup. BOLA: Authenticated but no ownership check; returns extended order fields.",
        "operationId": "getOrderStatusLegacy",
        "deprecated": true,
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID of the user"
          },
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID of the order to retrieve"
          },
          {
            "name": "legacy_token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Deprecated legacy token parameter (accepted but not validated)"
          }
        ],
        "responses": {
          "200": {
            "description": "Legacy order status response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyOrderStatus"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Order not found"
          }
        }
      }
    },
...
    "/api/users/{userId}/orders/{orderId}/apply-coupon": {
      "post": {
        "tags": [
          "Orders"
        ],
        "summary": "Apply coupon to order",
        "description": "Applies a coupon code. Demo vulnerability: coupon codes can be guessed.",
        "operationId": "applyCouponToOrder",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID of the order owner"
          },
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID of the order"
          },
          {
            "name": "coupon_code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Coupon code to apply"
          }
        ],
        "responses": {
          "200": {
            "description": "Order with coupon applied",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Coupon invalid or order state disallows change"
          },
          "403": {
            "description": "Not authorized for this order"
          },
          "404": {
            "description": "User or Order not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Address": {
        "type": "object",
        "required": [
          "address_id",
          "user_id",
          "street",
          "city",
          "country",
          "zip_code"
        ],
        "properties": {
          "address_id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier for the address",
            "readOnly": true
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "description": "Foreign key to User",
            "readOnly": true
          },
          "street": {
            "type": "string",
            "description": "Street address"
          },
          "city": {
            "type": "string",
            "description": "City"
          },
          "country": {
            "type": "string",
            "description": "Country"
          },
          "zip_code": {
            "type": "string",
            "description": "Zip code"
          },
          "is_default": {
            "type": "boolean",
            "default": false,
            "description": "Whether this is the default address for the user"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp of address creation",
            "readOnly": true
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp of last address update",
            "readOnly": true
          },
          "is_protected": {
            "type": "boolean",
            "description": "Indicates if this address is protected from deletion or modification",
            "default": false
          }
        }
      },
...
      "Order": {
        "type": "object",
        "required": [
          "order_id",
          "user_id",
          "address_id",
          "credit_card_id",
          "total_amount",
          "status"
        ],
        "properties": {
          "order_id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier for the order",
            "readOnly": true
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "description": "Foreign key to User (owner of the order)",
            "readOnly": true
          },
          "address_id": {
            "type": "string",
            "format": "uuid",
            "description": "Foreign key to Address used for shipping"
          },
          "credit_card_id": {
            "type": "string",
            "format": "uuid",
            "description": "Foreign key to CreditCard used for payment"
          },
          "applied_coupon_id": {
            "type": "string",
            "format": "uuid",
            "description": "Identifier of the applied coupon, if any",
            "nullable": true
          },
          "applied_coupon_code": {
            "type": "string",
            "description": "Code of the applied coupon, if any",
            "nullable": true
          },
          "discount_amount": {
            "type": "number",
            "format": "float",
            "description": "Discount applied to the order",
            "default": 0.0
          },
          "total_amount": {
            "type": "number",
            "format": "float",
            "description": "Total amount of the order after discount",
            "readOnly": true
          },
          "status": {
            "type": "string",
            "description": "Status of the order (e.g., pending, paid, shipped, delivered)",
            "default": "pending"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp of order creation",
            "readOnly": true
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp of last order update",
            "readOnly": true
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            },
            "default": []
          },
          "credit_card_last_four": {
            "type": "string",
            "nullable": true,
            "description": "Last four digits of the credit card used for the order, if available.",
            "readOnly": true
          }
        }
      },
      "LegacyOrderStatus": {
        "type": "object",
        "properties": {
          "order_id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "address": {
            "$ref": "#/components/schemas/Address",
            "nullable": true
          },
          "credit_card_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "credit_card_last_four": {
            "type": "string",
            "nullable": true
          },
          "billing_email": {
            "type": "string",
            "format": "email",
            "nullable": true
          },
          "legacy_notes": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "OrderItem": {
        "type": "object",
        "required": [
          "order_item_id",
          "order_id",
          "product_id",
          "quantity",
          "price_at_purchase"
        ],
        "properties": {
          "order_item_id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier for the order item",
            "readOnly": true
          },
          "order_id": {
            "type": "string",
            "format": "uuid",
            "description": "Foreign key to Order",
            "readOnly": true
          },
          "product_id": {
            "type": "string",
            "format": "uuid",
            "description": "Foreign key to Product"
          },
          "quantity": {
            "type": "integer",
            "description": "Quantity of the product ordered"
          },
          "price_at_purchase": {
            "type": "number",
            "format": "float",
            "description": "Price of the product at the time of purchase",
            "readOnly": true
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}